spike: dogfooding tests with test runner
