+++ package.json
@@ -35,7 +35,8 @@
   "scripts": {
     "lint": "./node_modules/.bin/eslint --color src",
     "publish": "",
-    "postpublish": "PACKAGE_VERSION=$(cat package.json | grep \\\"version\\\" | head -1 | awk -F: '{ print $2 }' | sed 's/[\",]//g' | tr -d '[[:space:]]') && git tag v$PACKAGE_VERSION && git push upstream --tags"
+    "postpublish": "PACKAGE_VERSION=$(cat package.json | grep \\\"version\\\" | head -1 | awk -F: '{ print $2 }' | sed 's/[\",]//g' | tr -d '[[:space:]]') && git tag v$PACKAGE_VERSION && git push upstream --tags",
+    "test": "node ./bin/test.js"
   },
   "dependencies": {
     "ejs": "2.6.1",
